+++ internal/infrastructure/s3/s3_repository.go
@@ -3,8 +3,8 @@ package s3
 import (
 	"fmt"
 
-	"github.com/ryo-funaba/example_echo/internal/pkg/errorutil"
-	"github.com/ryo-funaba/example_echo/internal/pkg/logutil"
+	"github.com/ryo-funaba/example_echo/internal/utils/errorutil"
+	"github.com/ryo-funaba/example_echo/internal/utils/logutil"
 
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
