+++ internal/utils/enum/connection_type.go
